creation of the ProfileCard component and works on the design of this component
