+++ frontend/package.json
@@ -21,6 +21,7 @@
     "@material-ui/core": "^4.11.14",
     "@material-ui/icons": "^4.11.3",
     "@material-ui/styles": "^4.11.5",
+    "@mui/base": "^5.0.0-alpha.123",
     "@mui/icons-material": "^5.11.9",
     "@mui/joy": "^5.0.0-alpha.68",
     "@mui/material": "^5.11.10",
