--- frontend/src/components/Sidebar/SidebarOld.jsx
@@ -1,250 +1,0 @@
-import { useState, useEffect } from "react";
-import axios from "axios";
-import { Link } from "react-router-dom";
-import PropTypes from "prop-types";
-import { ProSidebar, Menu, MenuItem } from "react-pro-sidebar";
-import { Box, IconButton, Typography, useTheme } from "@mui/material";
-import "react-pro-sidebar/dist/css/styles.css";
-import AddCircleOutlineOutlinedIcon from "@mui/icons-material/AddCircleOutlineOutlined";
-import HomeOutlinedIcon from "@mui/icons-material/HomeOutlined";
-import MailOutlineIcon from "@mui/icons-material/MailOutline";
-import CalendarTodayOutlinedIcon from "@mui/icons-material/CalendarTodayOutlined";
-import PieChartOutlineOutlinedIcon from "@mui/icons-material/PieChartOutlineOutlined";
-import AlternateEmailOutlinedIcon from "@mui/icons-material/AlternateEmailOutlined";
-import MenuOutlinedIcon from "@mui/icons-material/MenuOutlined";
-import LogoutOutlinedIcon from "@mui/icons-material/LogoutOutlined";
-import { tokens } from "../../theme";
-// import defaultUserImage from "../../assets/user.png";
-
-function Item({ title, to, icon, selected, setSelected }) {
-  const theme = useTheme();
-  const colors = tokens(theme.palette.mode);
-  return (
-    <MenuItem
-      active={selected === title}
-      style={{
-        color: colors.grey[100],
-      }}
-      onClick={() => setSelected(title)}
-      icon={icon}
-    >
-      <Typography>{title}</Typography>
-      <Link to={to} />
-    </MenuItem>
-  );
-}
-
-Item.propTypes = {
-  title: PropTypes.string.isRequired,
-  to: PropTypes.string.isRequired,
-  icon: PropTypes.element.isRequired,
-  selected: PropTypes.string.isRequired,
-  setSelected: PropTypes.func.isRequired,
-};
-
-function Sidebar() {
-  const theme = useTheme();
-  const colors = tokens(theme.palette.mode);
-  const [isCollapsed, setIsCollapsed] = useState(false);
-  const [selected, setSelected] = useState("Dashboard");
-  const [currentUser, setCurrentUser] = useState({});
-
-  const getUser = () => {
-    axios
-      .get(`http://localhost:5000/users/1`)
-      .then((response) => response.data)
-      .then((data) => {
-        setCurrentUser(data);
-      });
-  };
-
-  useEffect(() => {
-    getUser();
-  }, []);
-
-  const [randomUserImage, setRandomUserImage] = useState("");
-
-  useEffect(() => {
-    async function fetchRandomUserImage() {
-      const response = await axios.get(
-        "https://randomuser.me/api/?gender=male"
-      );
-      const image = response.data.results[0].picture.large;
-      setRandomUserImage(image);
-    }
-    fetchRandomUserImage();
-  }, []);
-
-  return (
-    <Box
-      sx={{
-        "& .pro-sidebar-inner": {
-          background: `${colors.primary[400]} !important`,
-        },
-        "& .pro-icon-wrapper": {
-          backgroundColor: "transparent !important",
-        },
-        "& .pro-inner-item": {
-          padding: "5px 35px 5px 20px !important",
-        },
-        "& .pro-inner-item:hover": {
-          color: "#868dfb !important",
-        },
-        "& .pro-menu-item.active": {
-          color: "#6870fa !important",
-        },
-        "& .pro-sidebar > .pro-sidebar-inner > .pro-sidebar-layout ": {
-          height: "800px !important",
-        },
-      }}
-    >
-      <ProSidebar collapsed={isCollapsed}>
-        <Menu iconShape="square">
-          {/* LOGO AND MENU ICON */}
-          <MenuItem
-            onClick={() => setIsCollapsed(!isCollapsed)}
-            icon={isCollapsed ? <MenuOutlinedIcon /> : undefined}
-            style={{
-              margin: "10px 0 20px 0",
-              color: colors.grey[100],
-            }}
-          >
-            {!isCollapsed && (
-              <Box
-                display="flex"
-                justifyContent="space-between"
-                alignItems="center"
-                ml="15px"
-              >
-                <Typography variant="h4" color={colors.grey[100]}>
-                  DevHub Project
-                </Typography>
-                <IconButton onClick={() => setIsCollapsed(!isCollapsed)}>
-                  <MenuOutlinedIcon />
-                </IconButton>
-              </Box>
-            )}
-          </MenuItem>
-          {/* USER */}
-
-          {!isCollapsed && (
-            <Box mb="25px">
-              <Box display="flex" justifyContent="center" alignItems="center">
-                <Box key={currentUser.id}>
-                  {currentUser.user_image ? (
-                    <img
-                      style={{ width: "150px", borderRadius: "50%" }}
-                      src={currentUser.user_image}
-                      alt="user"
-                    />
-                  ) : (
-                    <img
-                      style={{ width: "150px", borderRadius: "50%" }}
-                      src={randomUserImage}
-                      alt="default user"
-                    />
-                  )}
-                </Box>
-              </Box>
-              <Box textAlign="center">
-                <Typography
-                  variant="h2"
-                  color={colors.grey[100]}
-                  fontWeight="bold"
-                  sx={{ m: "10px 0 0 0" }}
-                >
-                  <Box key={currentUser.id}>
-                    {`${currentUser.firstname} 
-                    ${currentUser.lastname}`}
-                  </Box>
-                </Typography>
-                <Typography variant="h5" color={colors.safran[500]}>
-                  <Box key={currentUser.id}>{currentUser.user_role}</Box>
-                </Typography>
-              </Box>
-            </Box>
-          )}
-          {/* MENU ITEMS  */}
-
-          <Box paddingLeft={isCollapsed ? undefined : "10%"}>
-            <Item
-              title="Dashboard"
-              to="/dashboard"
-              icon={<HomeOutlinedIcon />}
-              selected={selected}
-              setSelected={setSelected}
-            />
-
-            <Typography
-              variant="h6"
-              color={colors.grey[300]}
-              sx={{ m: "15px 0 5px 20px" }}
-            >
-              Data
-            </Typography>
-            <Item
-              title="Mailbox"
-              to="/mailbox"
-              icon={<MailOutlineIcon />}
-              selected={selected}
-              setSelected={setSelected}
-            />
-            <Item
-              title="Contacts "
-              to="/contact"
-              icon={<AlternateEmailOutlinedIcon />}
-              selected={selected}
-              setSelected={setSelected}
-            />
-
-            <Typography
-              variant="h6"
-              color={colors.grey[300]}
-              sx={{ m: "15px 0 5px 20px" }}
-            >
-              Pages
-            </Typography>
-            <Item
-              title="Add Project"
-              to="/add-project"
-              icon={<AddCircleOutlineOutlinedIcon />}
-              selected={selected}
-              setSelected={setSelected}
-            />
-            <Item
-              title="Calendar"
-              to="/calendar"
-              icon={<CalendarTodayOutlinedIcon />}
-              selected={selected}
-              setSelected={setSelected}
-            />
-
-            <Typography
-              variant="h6"
-              color={colors.grey[300]}
-              sx={{ m: "15px 0 5px 20px" }}
-            >
-              Charts
-            </Typography>
-            <Item
-              title="Project Progress"
-              to="/progress"
-              icon={<PieChartOutlineOutlinedIcon />}
-              selected={selected}
-              setSelected={setSelected}
-            />
-            <Item
-              title="LogOut"
-              to="/"
-              icon={<LogoutOutlinedIcon />}
-              selected={selected}
-              setSelected={setSelected}
-            />
-          </Box>
-        </Menu>
-      </ProSidebar>
-    </Box>
-  );
-}
-
-export default Sidebar;
