+++ frontend/src/pages/AddProject/AddProject.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { Box, useTheme } from "@mui/material";
-import AddTaskForm from "@components/TodoList/HandleSubmitTask";
+import AddTaskForm from "../../components/TodoList/HandleSubmitTask";
 import AddProjectForm from "../../components/TodoList/HandleSubmitProject";
 import { tokens } from "../../theme";
 import Header from "../../components/Header/Header";
